# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: my_customTools API
  description: |
    媒体转码工具集后端 API 文档。

    包含视频转码、图片处理、动图转换、文档转换等功能。

    ## 认证方式
    使用 JWT Bearer Token 认证，在请求头中添加：
    ```
    Authorization: Bearer <token>
    ```

    ## 响应格式
    所有接口统一返回格式：
    ```json
    {
      "code": 0,
      "msg": "ok",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: 开发环境
  - url: http://production-server/api
    description: 生产环境
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证
  schemas:
...
    ApiResponse:
      type: object
      required:
        - code
        - msg
        - data
      properties:
        code:
          type: integer
          description: 业务状态码，0 表示成功
          example: 0
        msg:
          type: string
          description: 响应消息
          example: "ok"
        data:
          type: object
          description: 响应数据
...
    TaskStatus:
      type: string
      enum: [waiting, uploading, processing, completed, failed, cancelled]
    TaskType:
      type: string
      enum: [video, img, anim, document]
    BaseTask:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/TaskType'
        userId:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        inputPath:
          type: string
        outputPath:
          type: string
        inputFormat:
          type: string
        outputFormat:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        config:
          type: object
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        errorMsg:
          type: string
...
    DocumentTaskSubtype:
      type: string
      enum: [word-to-pdf, excel-to-csv, excel-to-word, pdf-merge, pdf-split]
    DocumentTranscodeConfig:
      type: object
      properties:
        subtype:
          $ref: '#/components/schemas/DocumentTaskSubtype'
        wordToPdf:
          type: object
          properties:
            format:
              type: string
              enum: [pdf, pdfa]
            quality:
              type: string
              enum: [screen, print, high]
        excelToCsv:
          type: object
          properties:
            sheet:
              oneOf:
                - type: string
                - type: integer
            delimiter:
              type: string
              enum: [',', ';', '\t']
            encoding:
              type: string
              enum: [utf-8, gbk]
        excelToWord:
          type: object
          properties:
            sheet:
              oneOf:
                - type: string
                - type: integer
            template:
              type: string
        pdfMerge:
          type: object
          properties:
            inputPaths:
              type: array
              items:
                type: string
            outputName:
              type: string
        pdfSplit:
          type: object
          properties:
            mode:
              type: string
              enum: [page, range, all]
            pages:
              type: array
              items:
                type: integer
            ranges:
              type: array
              items:
                type: string

    # ========== 上传相关 ==========
...
paths:
...
  /download/batch:
    post:
      tags: [Download]
      summary: 批量打包下载
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - taskIds
              properties:
                taskIds:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 50
      responses:
        '200':
          description: 下载文件流
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '400':
          description: 参数错误
  /download/{taskId}:
    get:
      tags: [Download]
      summary: 单文件下载
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 下载文件流
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: 任务未完成
        '404':
          description: 任务不存在或无权限

  # ==================== Config ====================
...
  /document/create:
    post:
      tags: [Document]
      summary: 创建文档转换任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subtype
              properties:
                subtype:
                  $ref: '#/components/schemas/DocumentTaskSubtype'
                fileName:
                  type: string
                inputPath:
                  type: string
                inputPaths:
                  type: array
                  items:
                    type: string
                inputFormat:
                  type: string
                config:
                  $ref: '#/components/schemas/DocumentTranscodeConfig'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '400':
          description: 参数错误
  /document/merge-pdf:
    post:
      tags: [Document]
      summary: 合并 PDF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inputPaths
              properties:
                inputPaths:
                  type: array
                  items:
                    type: string
                  minItems: 2
                outputName:
                  type: string
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '400':
          description: 参数错误
  /document/split-pdf:
    post:
      tags: [Document]
      summary: 拆分 PDF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inputPath
                - mode
              properties:
                inputPath:
                  type: string
                mode:
                  type: string
                  enum: [page, range, all]
                pages:
                  type: array
                  items:
                    type: integer
                ranges:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '400':
          description: 参数错误
  /document/word-to-pdf:
    post:
      tags: [Document]
      summary: Word 转 PDF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inputPath
              properties:
                inputPath:
                  type: string
                format:
                  type: string
                  enum: [pdf, pdfa]
                  default: pdf
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '400':
          description: 参数错误
  /document/excel-to-csv:
    post:
      tags: [Document]
      summary: Excel 转 CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inputPath
              properties:
                inputPath:
                  type: string
                sheet:
                  oneOf:
                    - type: string
                    - type: integer
                delimiter:
                  type: string
                  enum: [',', ';', '\t']
                  default: ','
                encoding:
                  type: string
                  enum: [utf-8, gbk]
                  default: utf-8
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '400':
          description: 参数错误
  /document/excel-to-word:
    post:
      tags: [Document]
      summary: Excel 转 Word
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inputPath
              properties:
                inputPath:
                  type: string
                sheet:
                  oneOf:
                    - type: string
                    - type: integer
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '400':
          description: 参数错误